v5.0/backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
pyarrow>=17.0.0
pandas>=2.2.0
numpy>=1.26.0
python-dateutil>=2.9.0
orjson>=3.10.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
aiofiles>=24.1.0
aiosqlite>=0.20.0
PyJWT>=2.8.0
bcrypt>=4.0.0
stripe>=8.0.0
